location + save technics

+++ domain/src/main/java/com/example/dronevision/domain/model/Coordinates.kt
@@ -1,6 +1,6 @@
 package com.example.dronevision.domain.model
 
 data class Coordinates(
-    val x: Float,
-    val y: Float
+    val x: Double,
+    val y: Double
 )
